refactor(chore): move base folder from domain/ to common/ for shared access
