+++ src/main/java/com/youssef/GridPulse/domain/device/repository/DeviceHistoryRepository.java
@@ -1,6 +1,6 @@
 package com.youssef.GridPulse.domain.device.repository;
 
-import com.youssef.GridPulse.domain.base.BaseHistoryRepository;
+import com.youssef.GridPulse.common.base.BaseHistoryRepository;
 import com.youssef.GridPulse.domain.device.entity.DeviceHistory;
 
 import java.util.UUID;
