+++ src/main/java/com/youssef/GridPulse/domain/identity/user/entity/UserHistory.java
@@ -1,6 +1,6 @@
 package com.youssef.GridPulse.domain.identity.user.entity;
 
-import com.youssef.GridPulse.domain.base.BaseHistoryEntity;
+import com.youssef.GridPulse.common.base.BaseHistoryEntity;
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import lombok.*;
