+++ src/main/java/com/youssef/GridPulse/domain/identity/user/repository/UserHistoryRepository.java
@@ -1,6 +1,6 @@
 package com.youssef.GridPulse.domain.identity.user.repository;
 
-import com.youssef.GridPulse.domain.base.BaseHistoryRepository;
+import com.youssef.GridPulse.common.base.BaseHistoryRepository;
 import com.youssef.GridPulse.domain.identity.user.entity.UserHistory;
 import org.springframework.stereotype.Repository;
