+++ src/test/java/com/youssef/GridPulse/domain/base/BaseMapperTest.java
@@ -1,5 +1,8 @@
 package com.youssef.GridPulse.domain.base;
 
+import com.youssef.GridPulse.common.base.BaseEntity;
+import com.youssef.GridPulse.common.base.BaseHistoryEntity;
+import com.youssef.GridPulse.common.base.BaseMapper;
 import org.junit.jupiter.api.*;
 
 import java.time.Instant;
