+++ src/test/java/com/youssef/GridPulse/domain/inverter/repository/InverterHistoryRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.youssef.GridPulse.domain.inverter.repository;
 
-import com.youssef.GridPulse.domain.base.BaseHistoryRepository;
+import com.youssef.GridPulse.common.base.BaseHistoryRepository;
 import com.youssef.GridPulse.domain.base.BaseHistoryRepositoryTest;
 import com.youssef.GridPulse.domain.inverter.entity.InverterHistory;
 import org.springframework.beans.factory.annotation.Autowired;
